+++ paper/src/main/java/com/badbones69/crazycrates/core/impl/Settings.java
@@ -1,8 +1,8 @@
 package com.badbones69.crazycrates.core.impl;
 
 import ch.jalu.configme.SettingsManager;
-import com.badbones69.crazycrates.core.config.ConfigManager;
-import com.badbones69.crazycrates.core.config.impl.ConfigKeys;
+import com.badbones69.common.config.ConfigManager;
+import com.badbones69.common.config.impl.ConfigKeys;
 import us.crazycrew.crazycrates.platform.ISettings;
 import java.util.List;