+++ changelog.md
@@ -1,4 +1,61 @@
 ## Changes 🔨
+### New API
+CrazyCrates API has been bumped to 0.9.0
+
+#### A new class has been added.
+Key Manager class, which allows you to check Items that players are holding.
+
+You can fetch this by doing CratesProvider#get()#getKeyManager()
+
+```java
+/**
+ * A class for the KeyManager, It handles checking physical keys, and anything else we might need soon.
+ *
+ * @author Ryder Belserion
+ * @version 0.9.0
+ * @since 0.9.0
+ */
+public abstract class KeyManager<I> {
+
+    /**
+     * The default constructor for {@link KeyManager<I>}.
+     *
+     * @author Ryder Belserion
+     * @since 0.9.0
+     */
+    public KeyManager() {}
+
+    /**
+     * Checks if the item is a valid key.
+     *
+     * @param item the ItemStack
+     * @return true or false
+     * @since 0.9.0
+     */
+    public abstract boolean isKey(@NotNull final I item);
+
+    /**
+     * Checks if an item matches another item.
+     *
+     * @param item the initial ItemStack
+     * @param comparing the ItemStack to compare
+     * @return true or false
+     * @since 0.9.0
+     */
+    public abstract boolean isMatchingKey(@NotNull final I item, @NotNull final I comparing);
+
+    /**
+     * Gets the name of the crate the key belongs to.
+     *
+     * @param item the ItemStack
+     * @return the name of the crate
+     * @since 0.9.0
+     */
+    public abstract String getKey(@NotNull final I item);
+
+}
+```
+
 ### Configuration changes
 - `ChestLines` has been renamed to `Rows` in the crate config files, [5eec53b](https://github.com/Crazy-Crew/CrazyCrates/commit/5eec53b)
   - All existing configurations should continue to work, You can run `/crazyrates migrate -mt CratesDeprecated` to change this, or simply use Find and Replace.
@@ -35,6 +92,7 @@
   - [0e139f6](https://github.com/Crazy-Crew/CrazyCrates/commit/0e139f6)
 - Fixed an issue with potion color not applying with the new items section.
 - Fixed an issue with the shield banners using the new item format.
+- Fixed an issue with the tier chance per prize not properly showing.
 - Fixed crates debug command.
 
 As always, Report 🐛 to https://github.com/Crazy-Crew/CrazyCrates/issues